+++ internal/server/handlers.go
@@ -91,80 +91,68 @@ func (s *Server) run(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if runBuiltins {
-		s.logger.Debug().Msg("starting builtin run")
+		s.logger.Debug().Msg("builtin start")
 		s.builtins.Run(id)
-		s.logger.Debug().Msg("builtin run done")
-		s.logger.Debug().Msg("calling builtin flush")
 		builtinMetrics := s.builtins.Flush(id)
-		s.logger.Debug().Msg("builtin flush done")
 		for metricName, metric := range *builtinMetrics {
 			metrics[metricName] = metric
 		}
+		s.logger.Debug().Msg("builtin done")
 	}
 
 	if runPlugins {
 		// NOTE: errors are ignored from plugins.Run
 		//       1. errors are already logged by Run
 		//       2. do not expose execution state to callers
-		s.logger.Debug().Msg("starting plugin run")
+		s.logger.Debug().Msg("plugin start")
 		s.plugins.Run(id)
-		s.logger.Debug().Msg("plugin run done")
-		s.logger.Debug().Msg("calling plugin flush")
 		pluginMetrics := s.plugins.Flush(id)
-		s.logger.Debug().Msg("plugin flush done")
 		for metricName, metric := range *pluginMetrics {
 			metrics[metricName] = metric
 		}
+		s.logger.Debug().Msg("plugin done")
 	}
 
 	if flushReceiver {
-		s.logger.Debug().Msg("calling receiver flush")
+		s.logger.Debug().Msg("receiver start")
 		receiverMetrics := receiver.Flush()
-		s.logger.Debug().Msg("receiver flush done")
 		for metricName, metric := range *receiverMetrics {
 			metrics[metricName] = metric
 		}
+		s.logger.Debug().Msg("receiver done")
 	}
 
 	if flushStatsd {
 		if s.statsdSvr != nil {
-			s.logger.Debug().Msg("calling statsd flush")
+			s.logger.Debug().Msg("statsd start")
 			statsdMetrics := s.statsdSvr.Flush()
-			s.logger.Debug().Msg("statsd flush done")
 			if statsdMetrics != nil {
 				pfx := viper.GetString(config.KeyStatsdHostCategory)
 				for metricName, metric := range *statsdMetrics {
 					metrics[pfx+config.MetricNameSeparator+metricName] = metric
 				}
 			}
+			s.logger.Debug().Msg("statsd done")
 		}
 	}
 
 	if flushProm {
-		s.logger.Debug().Msg("calling prom flush")
+		s.logger.Debug().Msg("prom start")
 		promMetrics := promrecv.Flush()
-		s.logger.Debug().Msg("prom flush done")
 		for metricName, metric := range *promMetrics {
 			metrics[metricName] = metric
 		}
+		s.logger.Debug().Msg("prom done")
 	}
 
-	s.logger.Debug().Msg("update lastMetrics")
 	lastMetrics.metrics = metrics
 	lastMetrics.ts = time.Now()
-	s.logger.Debug().Msg("update lastMetrics done")
 
-	s.logger.Debug().Msg("calling enable new metrics")
 	if err := s.check.EnableNewMetrics(&metrics); err != nil {
 		s.logger.Warn().Err(err).Msg("unable to update check metrics")
 	}
-	s.logger.Debug().Msg("enable new metrics done")
 
-	s.logger.Debug().Msg("encoding metrics")
 	s.encodeResponse(&metrics, w, r)
-	s.logger.Debug().Msg("encoding done")
-
-	s.logger.Info().Msgf("sent %d metrics", len(metrics))
 }
 
 // encodeResponse takes care of encoding the response to an HTTP request for metrics.
@@ -230,6 +218,8 @@ func (s *Server) encodeResponse(m *cgm.Metrics, w http.ResponseWriter, r *http.R
 			Msg("writing metrics to response")
 		return
 	}
+
+	s.logger.Info().Msgf("sent %d metrics", len(*m))
 
 	dumpDir := viper.GetString(config.KeyDebugDumpMetrics)
 	if dumpDir != "" {